Change type of function parameter

+++ include/mbedtls/cipher.h
@@ -433,9 +433,9 @@ const mbedtls_cipher_info_t *mbedtls_cipher_info_from_type(const mbedtls_cipher_
  *                      given \p cipher_id.
  * \return              \c NULL if the associated cipher information is not found.
  */
-const mbedtls_cipher_info_t *mbedtls_cipher_info_from_values(const mbedtls_cipher_id_t cipher_id,
-                                                             int key_bitlen,
-                                                             const mbedtls_cipher_mode_t mode);
+const mbedtls_cipher_info_t *mbedtls_cipher_info_from_values(const  mbedtls_cipher_id_t cipher_id,
+                                                             size_t key_bitlen,
+                                                             const  mbedtls_cipher_mode_t mode);
 
 /**
  * \brief               Retrieve the identifier for a cipher info structure.
